Fitting semiparametric AUC regression model adjusting for categorical covariates

sAUC is used to fit semiparametric AUC regression model specified by giving a formula object of response and covariates and a separate argument of treatment group. It will convert variables other than response into factors, estimate model parameters, and display results.

sAUC(formula = FALSE, treatment_group = FALSE, data = FALSE)

Arguments

formula
A formula object with response and covariates such as response ~ x1 + x2

treatment_group
A treatment group for which a comparision is to be made

data
A dataframe that contains variables needed for the analysis.

Value

A list of model summary, coefficients, AUC details, and session information.

Examples

ds <- NULL
for (x1 in 0:1){
 for (x2 in 0:2){
   for (x3 in 0:2){
     for (group in 0:1){
         response <- round(rnorm(n = 100, mean = 0, sd = 1),4)
         column <- cbind(x1,x2,x3, group, response)
         ds <- as.data.frame(rbind(ds, column))
   }
  }
 }
}
ds[,c("x1", "x2", "x3", "group")] <- lapply(ds[,c("x1", "x2", "x3", "group")],
                                          function(x) factor(x))

sAUC(formula = response ~ x1 + x2 + x3, treatment_group = "group", data = ds)

#>  

#> Data are being analyzed. Please, be patient.
#> 

#> The model is:  logit [ p ( Y_0  >  Y_1 ) ]  =  beta_0 +  beta_1*x11 + beta_2*x21 + beta_3*x22 + beta_4*x31 + beta_5*x32  
#> 
#> Model Summary
#> 

#> $`Model summary`
#>             Coefficients Std. Error    2.5%  97.5% Pr(>|z|)
#> (Intercept)      -0.0795     0.0954 -0.2666 0.1076   0.4050
#> x11               0.0517     0.0780 -0.1011 0.2045   0.5069
#> x21              -0.0731     0.0955 -0.2602 0.1140   0.4439
#> x22              -0.0871     0.0954 -0.2742 0.0999   0.3612
#> x31               0.0516     0.0955 -0.1355 0.2388   0.5885
#> x32               0.0450     0.0955 -0.1421 0.2321   0.6373
#> 
#> $Coefficients
#>                    [,1]
#> (Intercept) -0.07948193
#> x11          0.05173681
#> x21         -0.07309892
#> x22         -0.08713200
#> x31          0.05164917
#> x32          0.04502029
#> 
#> $`AUC details`
#>     auchat     finvhat logitauchat v_finv_auchat var_logitauchat
#> 1  0.49710 -0.01160013 -0.01160013    0.02729318      0.02729318
#> 2  0.51090  0.04360691  0.04360691    0.02722852      0.02722852
#> 3  0.42600 -0.29819005 -0.29819005    0.02761797      0.02761797
#> 4  0.50820  0.03280294  0.03280294    0.02702084      0.02702084
#> 5  0.44580 -0.21765521 -0.21765521    0.02731529      0.02731529
#> 6  0.45135 -0.19521762 -0.19521762    0.02758157      0.02758157
#> 7  0.44630 -0.21563165 -0.21563165    0.02721569      0.02721569
#> 8  0.52440  0.09767759  0.09767759    0.02725487      0.02725487
#> 9  0.42050 -0.32072118 -0.32072118    0.02778033      0.02778033
#> 10 0.53360  0.13460286  0.13460286    0.02723632      0.02723632
#> 11 0.52695  0.10790458  0.10790458    0.02720786      0.02720786
#> 12 0.46480 -0.14103330 -0.14103330    0.02745105      0.02745105
#> 13 0.46225 -0.15128790 -0.15128790    0.02730603      0.02730603
#> 14 0.52730  0.10930871  0.10930871    0.02735533      0.02735533
#> 15 0.53205  0.12837602  0.12837602    0.02727305      0.02727305
#> 16 0.43700 -0.25334643 -0.25334643    0.02758205      0.02758205
#> 17 0.51630  0.06522311  0.06522311    0.02706722      0.02706722
#> 18 0.43150 -0.27573380 -0.27573380    0.02753267      0.02753267
#> 
#> $`Session information`
#> R version 3.4.2 (2017-09-28)
#> Platform: x86_64-w64-mingw32/x64 (64-bit)
#> Running under: Windows 7 x64 (build 7601) Service Pack 1
#> 
#> Matrix products: default
#> 
#> locale:
#> [1] LC_COLLATE=English_United States.1252 
#> [2] LC_CTYPE=English_United States.1252   
#> [3] LC_MONETARY=English_United States.1252
#> [4] LC_NUMERIC=C                          
#> [5] LC_TIME=English_United States.1252    
#> 
#> attached base packages:
#> [1] stats     graphics  grDevices utils     datasets  methods   base     
#> 
#> other attached packages:
#> [1] sAUC_0.0.1.9
#> 
#> loaded via a namespace (and not attached):
#>  [1] Rcpp_0.12.13         compiler_3.4.2       BiocInstaller_1.26.1
#>  [4] tools_3.4.2          highlight_0.4.7.2    testthat_1.0.2.9000 
#>  [7] digest_0.6.12        pkgbuild_0.0.0.9000  pkgload_0.0.0.9000  
#> [10] evaluate_0.10.1      memoise_1.1.0        tibble_1.3.4        
#> [13] debugme_1.0.2        lattice_0.20-35      rlang_0.1.2         
#> [16] Matrix_1.2-11        rstudioapi_0.7       commonmark_1.4      
#> [19] curl_3.0             yaml_2.1.14          pkgdown_0.1.0.9000  
#> [22] withr_2.0.0          httr_1.3.1           stringr_1.2.0       
#> [25] roxygen2_6.0.1.9000  xml2_1.1.1           knitr_1.17          
#> [28] desc_1.1.1           rprojroot_1.2        grid_3.4.2          
#> [31] R6_2.2.2             processx_2.0.0.1     rmarkdown_1.6       
#> [34] purrr_0.2.3          callr_1.0.0.9000     magrittr_1.5        
#> [37] whisker_0.3-2        backports_1.1.1      htmltools_0.3.6     
#> [40] MASS_7.3-47          usethis_0.1.0        testit_0.7          
#> [43] assertthat_0.2.0     stringi_1.1.5        crayon_1.3.4        
#> 
#> $`Matrix of unique X levels `
#>    x1 x2 x3
#> 1   0  0  0
#> 2   1  0  0
#> 3   0  1  0
#> 4   1  1  0
#> 5   0  2  0
#> 6   1  2  0
#> 7   0  0  1
#> 8   1  0  1
#> 9   0  1  1
#> 10  1  1  1
#> 11  0  2  1
#> 12  1  2  1
#> 13  0  0  2
#> 14  1  0  2
#> 15  0  1  2
#> 16  1  1  2
#> 17  0  2  2
#> 18  1  2  2
#> 
#> $`Design matrix`
#>    (Intercept) x11 x21 x22 x31 x32
#> 1            1   0   0   0   0   0
#> 2            1   1   0   0   0   0
#> 3            1   0   1   0   0   0
#> 4            1   1   1   0   0   0
#> 5            1   0   0   1   0   0
#> 6            1   1   0   1   0   0
#> 7            1   0   0   0   1   0
#> 8            1   1   0   0   1   0
#> 9            1   0   1   0   1   0
#> 10           1   1   1   0   1   0
#> 11           1   0   0   1   1   0
#> 12           1   1   0   1   1   0
#> 13           1   0   0   0   0   1
#> 14           1   1   0   0   0   1
#> 15           1   0   1   0   0   1
#> 16           1   1   1   0   0   1
#> 17           1   0   0   1   0   1
#> 18           1   1   0   1   0   1
#> attr(,"assign")
#> [1] 0 1 2 2 3 3
#> attr(,"contrasts")
#> attr(,"contrasts")$x1
#> [1] "contr.treatment"
#> 
#> attr(,"contrasts")$x2
#> [1] "contr.treatment"
#> 
#> attr(,"contrasts")$x3
#> [1] "contr.treatment"
#> 
#> 
#> $model_formula
#> [1] "logit [ p ( Y 0  >  Y 1 ) ] \n\n"
#> 
#> $input_covariates
#> [1] "x1" "x2" "x3"
#> 
#> $input_response
#> [1] "response"
#>